requirements.txt
python-dotenv==0.8.2
SQLAlchemy==1.2.6
pymysql==0.8.0